Start to convert jactr tests
The tests from the org.jactr.tests bundle have been moved to the
respective modules, mostly core and io. Since the tests for core
depended on io, but the core artifact does not depend on io, I started
to re-write the tests to work without io. This work is not yet complete
for the test classes in core/src/main/java/org/jactr/modules/** that
still have compiler errors because they still require removal of io
dependencies. Some of the already converted tests fail and need to be
compared to the test results for github.com/amharrison/jactr.